+++ src/main/java/racinggame/domain/RacingCar.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 import java.util.*;
 import java.util.stream.Collectors;
